requirements.txt
asgiref==3.5.2
cffi==1.15.0
cryptography==37.0.2
Django==4.0.5
django-rest-knox==4.2.0
djangorestframework==3.13.1
pycparser==2.21
pytz==2022.1
sqlparse==0.4.2
